installed libc6-dev-i386 for 32 bit machine

+++ 0x00-hello_world/6-size.c
@@ -13,11 +13,11 @@ int main(void)
 	long long ll;
 	float f;
 
-	printf("Size of a char: %lu\n", (unsigned long)sizeof(c));
-	printf("Size of an int: %lu\n", (unsigned long)sizeof(i));
-	printf("Size of a long int: %lu\n", (unsigned long)sizeof(l));
-	printf("Size of a long long intr: %lu\n", (unsigned long)sizeof(ll));
-	printf("Size of a float: %lu\n", (unsigned long)sizeof(f));
+	printf("Size of a char: %lu\n", sizeof(c));
+	printf("Size of an int: %lu\n", sizeof(i));
+	printf("Size of a long int: %lu\n", sizeof(l));
+	printf("Size of a long long intr: %lu\n", sizeof(ll));
+	printf("Size of a float: %lu\n", sizeof(f));
 
 	return (0);
 }
